Validate config and split plugin parameter helpers
